test(EuiDataGridCell): move EuiDataGrid pagination jest tests to a separate file
