+++ packages/eui/src/components/datagrid/data_grid.test.tsx
@@ -468,228 +468,6 @@ describe('EuiDataGrid', () => {
     />
   );
 
-  describe('pagination', () => {
-    it('renders', () => {
-      const component = mount(
-        <EuiDataGrid
-          aria-label="test grid"
-          columns={[{ id: 'Column' }]}
-          columnVisibility={{
-            visibleColumns: ['Column'],
-            setVisibleColumns: () => {},
-          }}
-          rowCount={10}
-          renderCellValue={renderCellRowAsValue}
-          pagination={{
-            pageIndex: 1,
-            pageSize: 6,
-            pageSizeOptions: [3, 6, 10],
-            onChangePage: () => {},
-            onChangeItemsPerPage: () => {},
-          }}
-        />
-      );
-
-      expect(component.find('EuiTablePagination').render()).toMatchSnapshot();
-    });
-
-    describe('page navigation', () => {
-      it('next button pages through content', () => {
-        const component = mount(
-          <EuiDataGrid
-            aria-label="test grid"
-            columns={[{ id: 'Column' }]}
-            columnVisibility={{
-              visibleColumns: ['Column'],
-              setVisibleColumns: () => {},
-            }}
-            rowCount={8}
-            renderCellValue={renderCellRowAsValue}
-            pagination={{
-              pageIndex: 0,
-              pageSize: 3,
-              pageSizeOptions: [3, 6, 10],
-              onChangePage: jest.fn((pageIndex) => {
-                const pagination = component.props().pagination;
-                component.setProps({
-                  pagination: { ...pagination, pageIndex },
-                });
-              }),
-              onChangeItemsPerPage: jest.fn(),
-            }}
-          />
-        );
-
-        expect(extractGridData(component)).toEqual([
-          ['Column'],
-          ['0'],
-          ['1'],
-          ['2'],
-        ]);
-
-        findTestSubject(component, 'pagination-button-next').simulate('click');
-
-        expect(component.props().pagination.onChangePage).toHaveBeenCalledTimes(
-          1
-        );
-        const firstCallPageIndex =
-          component.props().pagination.onChangePage.mock.calls[0][0];
-        expect(firstCallPageIndex).toBe(1);
-
-        expect(extractGridData(component)).toEqual([
-          ['Column'],
-          ['3'],
-          ['4'],
-          ['5'],
-        ]);
-
-        findTestSubject(component, 'pagination-button-next').simulate('click');
-
-        expect(component.props().pagination.onChangePage).toHaveBeenCalledTimes(
-          2
-        );
-        const secondCallPageIndex =
-          component.props().pagination.onChangePage.mock.calls[1][0];
-        expect(secondCallPageIndex).toBe(2);
-
-        expect(extractGridData(component)).toEqual([['Column'], ['6'], ['7']]);
-      });
-
-      it('pages are navigable through page links', () => {
-        const component = mount(
-          <EuiDataGrid
-            aria-label="test grid"
-            columns={[{ id: 'Column' }]}
-            columnVisibility={{
-              visibleColumns: ['Column'],
-              setVisibleColumns: () => {},
-            }}
-            rowCount={8}
-            renderCellValue={renderCellRowAsValue}
-            pagination={{
-              pageIndex: 0,
-              pageSize: 3,
-              pageSizeOptions: [3, 6, 10],
-              onChangePage: jest.fn((pageIndex) => {
-                const pagination = component.props().pagination;
-                component.setProps({
-                  pagination: { ...pagination, pageIndex },
-                });
-              }),
-              onChangeItemsPerPage: jest.fn(),
-            }}
-          />
-        );
-
-        expect(extractGridData(component)).toEqual([
-          ['Column'],
-          ['0'],
-          ['1'],
-          ['2'],
-        ]);
-
-        // goto page 3
-        findTestSubject(component, 'pagination-button-2').simulate('click');
-
-        expect(component.props().pagination.onChangePage).toHaveBeenCalledTimes(
-          1
-        );
-        const firstCallPageIndex =
-          component.props().pagination.onChangePage.mock.calls[0][0];
-        expect(firstCallPageIndex).toBe(2);
-
-        expect(extractGridData(component)).toEqual([['Column'], ['6'], ['7']]);
-
-        // goto page 2
-        findTestSubject(component, 'pagination-button-1').simulate('click');
-
-        expect(component.props().pagination.onChangePage).toHaveBeenCalledTimes(
-          2
-        );
-        const secondCallPageIndex =
-          component.props().pagination.onChangePage.mock.calls[1][0];
-        expect(secondCallPageIndex).toBe(1);
-
-        expect(extractGridData(component)).toEqual([
-          ['Column'],
-          ['3'],
-          ['4'],
-          ['5'],
-        ]);
-      });
-    });
-
-    it('changes the page size', () => {
-      const component = mount(
-        <EuiDataGrid
-          aria-label="test grid"
-          columns={[{ id: 'Column' }]}
-          columnVisibility={{
-            visibleColumns: ['Column'],
-            setVisibleColumns: () => {},
-          }}
-          rowCount={8}
-          renderCellValue={renderCellRowAsValue}
-          pagination={{
-            pageIndex: 0,
-            pageSize: 3,
-            pageSizeOptions: [3, 6, 10],
-            onChangePage: jest.fn(),
-            onChangeItemsPerPage: jest.fn((pageSize) => {
-              const pagination = component.props().pagination;
-              component.setProps({
-                pagination: { ...pagination, pageSize },
-              });
-            }),
-          }}
-        />
-      );
-
-      expect(extractGridData(component)).toEqual([
-        ['Column'],
-        ['0'],
-        ['1'],
-        ['2'],
-      ]);
-
-      act(() => {
-        findTestSubject(component, 'tablePaginationPopoverButton').simulate(
-          'click'
-        );
-      });
-
-      const rowButtons: NodeListOf<HTMLButtonElement> =
-        document.body.querySelectorAll('.euiContextMenuItem');
-      expect(
-        Array.prototype.map.call(
-          rowButtons,
-          (button: HTMLDivElement) => button.textContent || ''
-        )
-      ).toEqual(['3 rows', '6 rows', '10 rows']);
-
-      act(() => {
-        rowButtons[1].click();
-      });
-
-      expect(
-        component.props().pagination.onChangeItemsPerPage
-      ).toHaveBeenCalledTimes(1);
-      const firstCallPageIndex =
-        component.props().pagination.onChangeItemsPerPage.mock.calls[0][0];
-      expect(firstCallPageIndex).toBe(6);
-
-      expect(extractGridData(component)).toEqual([
-        ['Column'],
-        ['0'],
-        ['1'],
-        ['2'],
-        ['3'],
-        ['4'],
-        ['5'],
-      ]);
-    });
-  });
-
   describe('column sizing', () => {
     it('uses a columns initialWidth', () => {
       const component = mount(
